requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncmy==0.2.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4